feat(cli): expose docgen-cli config interfaces
