+++ packages/vue-docgen-cli/README.md
@@ -152,7 +152,7 @@ By default it will find the `Readme.md` sibling to the component files. Use this
 
 > type: `(file: string, config: DocgenCLIConfig) => string`, default: `(file, config) => path.resolve(config.outDir, file).replace(/\.\w+\$/, '.md')`
 
-Function returning the absolute path of the documentation markdown files. It [outFile](#outfile) is used, this config will be ignored.
+Function returning the absolute path of the documentation markdown files. If [outFile](#outfile) is used, this config will be ignored.
 
 #### watch
